update configs and forward
