+++ nemo/collections/nlp/models/machine_translation/mt_enc_dec_model.py
@@ -53,24 +53,11 @@ class MTEncDecModel(EncDecNLPModel):
 
         super().__init__(cfg=cfg, trainer=trainer)
 
+        # encoder should be TransformerEncoderNM (HF and Megatron encoders are work in progress)
         self.encoder = instantiate(cfg.encoder)
-        self.encoder_embedding = TransformerEmbedding(
-            vocab_size=self.encoder_vocab_size,
-            hidden_size=cfg.encoder.hidden_size,
-            max_sequence_length=cfg.encoder_embedding.max_sequence_length,
-            num_token_types=cfg.encoder_embedding.num_token_types,
-            embedding_dropout=cfg.encoder_embedding.embedding_dropout,
-            learn_positional_encodings=cfg.encoder_embedding.learn_positional_encodings,
-        )
+
+        # decoder should be TransformerDecoderNM (HF and Megatron decoders are work in progress)
         self.decoder = instantiate(cfg.decoder)
-        self.decoder_embedding = TransformerEmbedding(
-            vocab_size=self.decoder_vocab_size,
-            hidden_size=cfg.decoder.hidden_size,
-            max_sequence_length=cfg.decoder_embedding.max_sequence_length,
-            num_token_types=cfg.decoder_embedding.num_token_types,
-            embedding_dropout=cfg.decoder_embedding.embedding_dropout,
-            learn_positional_encodings=cfg.decoder_embedding.learn_positional_encodings,
-        )
 
         self.log_softmax = TokenClassifier(
             hidden_size=cfg.decoder.hidden_size,
@@ -128,11 +115,9 @@ class MTEncDecModel(EncDecNLPModel):
         Returns:
 
         """
-        src_embeddings = self.encoder_embedding(input_ids=src)
-        src_hiddens = self.encoder(src_embeddings, src_mask)
+        src_hiddens = self.encoder(src, src_mask)
         if tgt is not None:
-            tgt_embeddings = self.decoder_embedding(input_ids=tgt)
-            tgt_hiddens = self.decoder(tgt_embeddings, tgt_mask, src_hiddens, src_mask)
+            tgt_hiddens = self.decoder(tgt, tgt_mask, src_hiddens, src_mask)
             log_probs = self.log_softmax(hidden_states=tgt_hiddens)
         else:
             log_probs = None
